Propagate the test-pattern size thru bytesused
When reading test-pattern files from disk the size of the pattern must
for the most cases match the configured size of each plane. The
exception is compressed or otherwise special format indicated by the
plane being configured with a bytesperline set to 0.

In the special cases the pattern is read up-to the size of the file on
disk and queued to the device. The size of the test pattern are however
not propagated correctly in the bytesused filed of the buffer being
queued to the vdev.

This makes it impossible to queue meta buffers containing extendable
parameters, e.g. V4L2_META_FMT_RK_ISP1_EXT_PARAMS, using yavta. As the
RkISP1 driver uses the information stored in bytesused to figure out how
much valid data there is in the buffer.

Fix this by updating the size of the test-pattern when reading it from
disk.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>