Use key_value to handle V4L2 field names
Extend the key_value helper with a function to retrieve the key
corresponding to a value, and use it to replace the manual
implementation of V4L2 field name lookup.

The name comparison needs to be made case-insensitive to avoid breaking
users, as yavta handles V4L2 field names in a case-insensitive way.

Suggested-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Reviewed-by: Sakari Ailus <sakari.ailus@linux.intel.com>